backend/week2/pyproject.toml
[project]
name = "week2"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
